Depends on #3595.
Improve CPU performance.

- `master`
```
contraction         :    CPU:   55.820 us   +/- 6.812 (min:   49.636 / max:   68.460) us     GPU-0:40316.032 us   +/-383.087 (min:39570.110 / max:40646.305) us
elementwise_binary  :    CPU:   31.403 us   +/- 3.674 (min:   27.876 / max:   37.964) us     GPU-0:  472.358 us   +/- 5.129 (min:  464.608 / max:  479.168) us
elementwise_trinary :    CPU:   37.470 us   +/- 5.565 (min:   31.486 / max:   45.147) us     GPU-0:  563.712 us   +/- 5.574 (min:  557.184 / max:  571.680) us
reduction           :    CPU:   39.862 us   +/- 6.660 (min:   35.085 / max:   52.668) us     GPU-0: 1016.749 us   +/- 6.731 (min: 1012.352 / max: 1030.048) us
```
- This PR (`2e654f7`)
```
contraction         :    CPU:   33.880 us   +/- 6.933 (min:   28.368 / max:   47.218) us     GPU-0:39861.286 us   +/-540.859 (min:38781.727 / max:40186.047) us
elementwise_binary  :    CPU:   20.511 us   +/- 2.704 (min:   17.621 / max:   25.388) us     GPU-0:  461.299 us   +/- 5.706 (min:  452.864 / max:  470.816) us
elementwise_trinary :    CPU:   23.350 us   +/- 2.867 (min:   20.951 / max:   28.692) us     GPU-0:  549.933 us   +/- 3.890 (min:  545.856 / max:  557.024) us
reduction           :    CPU:   23.770 us   +/- 4.193 (min:   20.538 / max:   31.900) us     GPU-0: 1011.245 us   +/- 8.741 (min: 1003.680 / max: 1027.936) us
```
Compared with the scripts in this repository: `examples/cutensor/*.py`.

cupy.sum with cuTENSOR accelerator (implemented in _try_reduction_routine) is also improved.
- `master
```
sum                 :    CPU:   57.215 us   +/- 9.320 (min:   55.607 / max:  913.702) us     GPU-0:  151.111 us   +/-10.814 (min:  147.168 / max:  987.008) us
```
- This PR (`ba1118a`)
```
sum                 :    CPU:   18.730 us   +/- 4.922 (min:   17.982 / max:  481.002) us     GPU-0:  114.798 us   +/- 4.809 (min:  111.904 / max:  560.192) us
```
Compared by running the following script with setting `CUPY_ACCELERATORS='cutensor'`.
```py
import cupy
import cupyx.time
x = cupy.arange(2 ** 24, dtype='float32')
perf = cupyx.time.repeat(cupy.sum, (x,))
print(perf)
```