@asi1024 2020-08-02
Now cpu-time overhead issue in cuTENSOR reduction accelerator was resolved.
```
# w/o cuTENSOR
sum                 :    CPU:   18.878 us   +/-23.512 (min:   16.116 / max:  641.480) us     GPU-0:11594.718 us   +/-25.859 (min:11567.744 / max:12233.152) us
# w/ cuTENSOR
sum                 :    CPU:   18.730 us   +/- 4.922 (min:   17.982 / max:  481.002) us     GPU-0:  114.798 us   +/- 4.809 (min:  111.904 / max:  560.192) us
```