@leofang 2020-08-02
Looks nice! @asi1024 could you note what makes the performance improvement? Also, 

> Now cpu-time overhead issue in cuTENSOR reduction accelerator was resolved.

By overhead you meant the large CPU time fluctuation? Why did it happen?