@asi1024 2020-08-03
@leofang This PR reduces the accessing of object attribute by annotating cython types.