@asi1024 2020-08-06
Hmm, I couldn't reproduce the build failure locally. (Perhaps it depends on Cython's version?) I made a fix.
Jenkins, test this please.